requirements.txt
Pillow>=10.3.0
requests>=2.32.0
imageio>=2.34.0
numpy>=1.26.0
PyYAML>=6.0.2
